Changing NRG in health benchmark to be archeture independant; also changing current inputs to be consistant with current implementation; output verified with serial execution; close #124

+++ omp-tasks/health/health.h
@@ -43,7 +43,7 @@ extern int sim_level;
 
 struct Patient {
    int id;
-   long seed;
+   int32_t seed;
    int time;
    int time_left;
    int hosps_visited;
@@ -68,10 +68,10 @@ struct Village {
    struct Patient *population;
    struct Hosp hosp;
    int level;
-   long  seed;
+   int32_t  seed;
 };
 
-float my_rand(long *seed);
+float my_rand(int32_t *seed);
 
 struct Patient *generate_patient(struct Village *village);
 void put_in_hosp(struct Hosp *hosp, struct Patient *patient);
@@ -92,7 +92,7 @@ float get_total_hosps(struct Village *village);
 struct Results get_results(struct Village *village);
 
 void read_input_data(char *filename);
-void allocate_village( struct Village **capital, struct Village *back, struct Village *next, int level, int vid);
+void allocate_village( struct Village **capital, struct Village *back, struct Village *next, int level, int32_t vid);
 void sim_village_main_par(struct Village *top);
 
 void sim_village_par(struct Village *village);
