+++ serial/health/health.h
@@ -41,7 +41,7 @@ struct Results {
 
 struct Patient {
    int id;
-   long seed;
+   int32_t seed;
    int time;
    int time_left;
    int hosps_visited;
@@ -66,10 +66,10 @@ struct Village {
    struct Patient *population;
    struct Hosp hosp;
    int level;
-   long  seed;
+   int32_t  seed;
 };
 
-float my_rand(long *seed);
+float my_rand(int32_t *seed);
 
 struct Patient *generate_patient(struct Village *village);
 void put_in_hosp(struct Hosp *hosp, struct Patient *patient);
@@ -94,7 +94,7 @@ void sim_village_par(struct Village *village);
 extern int sim_level;
 
 void read_input_data(char *filename);
-void allocate_village( struct Village **capital, struct Village *back, struct Village *next, int level, int vid);
+void allocate_village( struct Village **capital, struct Village *back, struct Village *next, int level, int32_t vid);
 void sim_village_main(struct Village *top);
 
 int check_village(struct Village *top);
